revert: workspace list accessibility changes to fix deploy blockers #82727 #82739
